Function hotshot_orchestrator::run_orchestrator

pub async fn run_orchestrator<KEY, ELECTION>(
    network_config: NetworkConfig<KEY, ELECTION>,
    url: Url
) -> Result<()>where
    KEY: SignatureKey + 'static + Serialize,
    ELECTION: ElectionConfig + 'static + Serialize,

Runs the orchestrator

Errors

This errors if tide disco runs into an issue during serving

Panics

This panics if unable to register the api with tide disco