Struct hotshot_orchestrator::OrchestratorState

pub(crate) struct OrchestratorState<KEY: SignatureKey> {

    pub(crate) latest_index: u16,
    pub(crate) tmp_latest_index: u16,
    pub(crate) config: NetworkConfig<KEY>,
    pub(crate) peer_pub_ready: bool,
    pub(crate) pub_posted: HashMap<Vec<u8>, (u64, bool)>,
    pub(crate) start: bool,
    pub(crate) nodes_connected: HashSet<PeerConfig<KEY>>,
    pub(crate) bench_results: BenchResults,
    pub(crate) nodes_post_results: u64,
    pub(crate) manual_start_allowed: bool,
    pub(crate) accepting_new_keys: bool,
    pub(crate) builders: Vec<Url>,
    pub(crate) fixed_stake_table: bool,

}

The state of the orchestrator

Fields
latest_index: u16

Tracks the latest node index we have generated a configuration for

tmp_latest_index: u16

Tracks the latest temporary index we have generated for init validator’s key pair

config: NetworkConfig<KEY>

The network configuration

peer_pub_ready: bool

Whether the network configuration has been updated with all the peer’s public keys/configs

pub_posted: HashMap<Vec<u8>, (u64, bool)>

A map from public keys to (node_index, is_da).

start: bool

Whether nodes should start their HotShot instances Will be set to true once all nodes post they are ready to start

nodes_connected: HashSet<PeerConfig<KEY>>

The total nodes that have posted they are ready to start

bench_results: BenchResults

The results of the benchmarks

nodes_post_results: u64

The number of nodes that have posted their results

manual_start_allowed: bool

Whether the orchestrator can be started manually

accepting_new_keys: bool

Whether we are still accepting new keys for registration

builders: Vec<Url>

Builder address pool

fixed_stake_table: bool

whether we are using a fixed stake table, disabling public key registration

Implementations

impl<KEY: SignatureKey + 'static> OrchestratorState<KEY>

pub fn new(network_config: NetworkConfig<KEY>) -> Self

create a new OrchestratorState

pub fn output_to_csv(&self)

Output the results to a csv file according to orchestrator state

impl<KEY> OrchestratorState<KEY>
where KEY: Serialize + Clone + SignatureKey + 'static,

pub(crate) fn register_unknown( &mut self, pubkey: &mut Vec<u8>, da_requested: bool, libp2p_address: Option<Multiaddr>, libp2p_public_key: Option<PeerId>, ) -> Result<(u64, bool), ServerError>

register a node with an unknown public key. this method should be used when we don’t have a fixed stake table

pub(crate) fn register_from_list( &mut self, pubkey: &mut Vec<u8>, da_requested: bool, libp2p_address: Option<Multiaddr>, libp2p_public_key: Option<PeerId>, ) -> Result<(u64, bool), ServerError>

register a node on the fixed stake table, which was loaded at startup