Trait hotshot_orchestrator::OrchestratorApi

pub trait OrchestratorApi<KEY: SignatureKey, ELECTION: ElectionConfig> {
    // Required methods
    fn post_identity(
        &mut self,
        libp2p_address: Option<Multiaddr>,
        libp2p_public_key: Option<PeerId>
    ) -> Result<u16, ServerError>;
    fn post_getconfig(
        &mut self,
        _node_index: u16
    ) -> Result<NetworkConfig<KEY, ELECTION>, ServerError>;
    fn get_tmp_node_index(&mut self) -> Result<u16, ServerError>;
    fn register_public_key(
        &mut self,
        node_index: u64,
        pubkey: &mut Vec<u8>
    ) -> Result<(), ServerError>;
    fn peer_pub_ready(&self) -> Result<bool, ServerError>;
    fn get_config_after_peer_collected(
        &self
    ) -> Result<NetworkConfig<KEY, ELECTION>, ServerError>;
    fn get_start(&self) -> Result<bool, ServerError>;
    fn post_run_results(
        &mut self,
        metrics: BenchResults
    ) -> Result<(), ServerError>;
    fn post_ready(&mut self) -> Result<(), ServerError>;
}

An api exposed by the orchestrator

Required Methods

fn post_identity( &mut self, libp2p_address: Option<Multiaddr>, libp2p_public_key: Option<PeerId> ) -> Result<u16, ServerError>

Post an identity to the orchestrator. Takes in optional arguments so others can identify us on the Libp2p network.

Errors

If we were unable to serve the request

fn post_getconfig( &mut self, _node_index: u16 ) -> Result<NetworkConfig<KEY, ELECTION>, ServerError>

post endpoint for each node’s config

Errors

if unable to serve

fn get_tmp_node_index(&mut self) -> Result<u16, ServerError>

get endpoint for the next available temporary node index

Errors

if unable to serve

fn register_public_key( &mut self, node_index: u64, pubkey: &mut Vec<u8> ) -> Result<(), ServerError>

post endpoint for each node’s public key

Errors

if unable to serve

fn peer_pub_ready(&self) -> Result<bool, ServerError>

post endpoint for whether or not all peers public keys are ready

Errors

if unable to serve

fn get_config_after_peer_collected( &self ) -> Result<NetworkConfig<KEY, ELECTION>, ServerError>

get endpoint for the network config after all peers public keys are collected

Errors

if unable to serve

fn get_start(&self) -> Result<bool, ServerError>

get endpoint for whether or not the run has started

Errors

if unable to serve

fn post_run_results(&mut self, metrics: BenchResults) -> Result<(), ServerError>

post endpoint for the results of the run

Errors

if unable to serve

fn post_ready(&mut self) -> Result<(), ServerError>

post endpoint for whether or not all nodes are ready

Errors

if unable to serve