Crate hotshot_orchestrator

Orchestrator for manipulating nodes and recording results during a run of HotShot tests

Modules

client

The orchestrator’s clients

config

Configuration for the orchestrator

Structs

OrchestratorState 🔒

The state of the orchestrator

Constants

ORCHESTRATOR_MAJOR_VERSION

Orchestrator is not, strictly speaking, bound to the network; it can have its own versioning. Orchestrator Version (major)

ORCHESTRATOR_MINOR_VERSION

Orchestrator Version (minor)

ORCHESTRATOR_VERSION

Orchestrator Version as a type-binding instance

Traits

OrchestratorApi

An api exposed by the orchestrator

Functions

define_api 🔒

Sets up all API routes

libp2p_generate_indexed_identity

Generate an keypair based on a seed and an index

run_orchestrator

Runs the orchestrator

Type Aliases

OrchestratorVersion

Orchestrator Version as a type